+++ src/image_test.rs
@@ -49,14 +49,14 @@ pub fn run_julia_set() {
     for x in 0..4000 {
         for y in 0..4000 {
             let pixel = img.get_pixel_mut(x, y);
-            let mut z = Complex::new((x as f64 - 2000.0) / 1000.0, (y as f64 - 2000.0) / 1000.0);
+            let mut z = Complex::new((x as f32 - 2000.0) / 1000.0, (y as f32 - 2000.0) / 1000.0);
             let c = Complex::new(-0.70176, -0.3842);
             let mut i = 0;
             while i < 70 && z.norm_sqr() < 2.0 * 2.0 {
                 z = z * z + c;
                 i += 1;
             }
-            let t = i as f64 / 70.0;
+            let t = i as f32 / 70.0;
             let color = Vec3::new(1.0, 1.0, 1.0) * (1.0 - t) + Vec3::new(0.5, 0.7, 1.0) * t;
             *pixel = image::Rgb([
                 (color.x * 255.99999) as u8,
