+++ src/ray_tracing.rs
@@ -51,7 +51,7 @@ pub fn run_ray_tracing() {
         (600, 600, 64, 16)
     };
 
-    let (world, background, cam) = cornell_box(image_width as f64 / image_height as f64);
+    let (world, background, cam) = cornell_box(image_width as f32 / image_height as f32);
 
     let mut img: RgbImage = ImageBuffer::new(image_width, image_height);
     let pbar = ProgressBar::new(image_width as u64);
@@ -72,12 +72,12 @@ pub fn run_ray_tracing() {
                 for y in 0..image_height {
                     let mut color = Vec3::new(0.0, 0.0, 0.0);
                     for _ in 0..samples_per_pixel {
-                        let u = (x as f64 + rng.gen::<f64>()) / (image_width as f64 - 1.0);
-                        let v = (y as f64 + rng.gen::<f64>()) / (image_height as f64 - 1.0);
+                        let u = (x as f32 + rng.gen::<f32>()) / (image_width as f32 - 1.0);
+                        let v = (y as f32 + rng.gen::<f32>()) / (image_height as f32 - 1.0);
                         let ray = cam.get_ray(u, v, &mut rng);
                         color += ray_color(&ray, &world, background, 50, &mut rng);
                     }
-                    color /= samples_per_pixel as f64;
+                    color /= samples_per_pixel as f32;
                     ans.color.push([
                         (num::clamp(color.x.sqrt(), 0.0, 0.99999) * 256.0) as u8,
                         (num::clamp(color.y.sqrt(), 0.0, 0.99999) * 256.0) as u8,
